+++ src/nvim/charset.h
@@ -2,13 +2,12 @@
 #define NVIM_CHARSET_H
 
 #include <stdbool.h>
+#include <stdint.h>
 
 #include "nvim/buffer_defs.h"
 #include "nvim/eval/typval_defs.h"
-#include "nvim/option_defs.h"
 #include "nvim/pos.h"
 #include "nvim/strings.h"
-#include "nvim/types.h"
 
 /// Return the folded-case equivalent of the given character
 ///
